Make the default binding mechanism to  be Masquerade
Since HCO's focus are legacy workloads where live migration is a common practice,
It is required to encourage  the use of `masquerade` binding over bridge.
However, the bridge binding is still available and it is even the default,
which sometimes makes the usage confusing for users.
That's why we would like masquerade to be our default binding mechanism.

to achieve that in a more "clean" and "appropriate" way, this PR exposes all kubevirt configuration  in the HCO-CR and then pass them to kubevirt configMap.
Thus making the HCO, which creates the configMap, be also responsible for giving the the default values, before kubevirt is even created.
I  believe this makes more sense.

After that, all that is left is to pass the default-binding-mechanism configuration a value of `masquerade` and get the desired behavior by default.
Signed-off-by: alonSadan <asadan@redhat.com>